refactor(mcp): Extract MCP module into submodules
Split src/mcp.rs (2,149 lines) into focused modules:

Structure:
src/mcp/
├── mod.rs          (57 lines - re-exports public API)
├── types.rs        (129 lines - JSON-RPC & MCP protocol types)
├── server.rs       (268 lines - McpServer core, initialization)
├── audit_mode.rs   (90 lines - audit mode state)
├── validation.rs   (211 lines - input validation helpers)
├── tools/
│   ├── mod.rs      (210 lines - dispatcher, Tool definitions)
│   ├── search.rs   (158 lines - semantic search)
│   ├── stats.rs    (71 lines - index statistics)
│   ├── call_graph.rs (68 lines - callers/callees)
│   ├── read.rs     (131 lines - file reading with context)
│   ├── notes.rs    (164 lines - add note)
│   └── audit.rs    (84 lines - audit mode toggle)
└── transports/
    ├── mod.rs      (9 lines - re-exports)
    ├── stdio.rs    (75 lines - stdin/stdout transport)
    └── http.rs     (559 lines - HTTP/SSE transport with tests)

All modules under 560 lines. Public API unchanged.
