+++ src/mcp/transports/mod.rs
@@ -1,0 +1,9 @@
+//! MCP transport implementations
+//!
+//! Transports provide different ways to communicate with the MCP server.
+
+mod http;
+mod stdio;
+
+pub use http::serve_http;
+pub use stdio::serve_stdio;
